+++ just_audio/lib/just_audio.dart
@@ -108,9 +108,13 @@ class AudioPlayer {
   bool _canUseNetworkResourcesForLiveStreamingWhilePaused = false;
   double _preferredPeakBitRate = 0;
   bool _playInterrupted = false;
+  bool _platformLoading = false;
   AndroidAudioAttributes? _androidAudioAttributes;
   final bool _androidApplyAudioAttributes;
   final bool _handleAudioSessionActivation;
+
+  /// Counts how many times [_setPlatformActive] is called.
+  int _activationCount = 0;
 
   /// Creates an [AudioPlayer].
   ///
@@ -150,7 +154,6 @@ class AudioPlayer {
         _audioLoadConfiguration = audioLoadConfiguration,
         _audioPipeline = audioPipeline ?? AudioPipeline() {
     _audioPipeline._setup(this);
-    _idlePlatform = _IdleAudioPlayer(id: _id, sequenceStream: sequenceStream);
     if (_audioLoadConfiguration?.darwinLoadControl != null) {
       _automaticallyWaitsToMinimizeStalling = _audioLoadConfiguration!
           .darwinLoadControl!.automaticallyWaitsToMinimizeStalling;
@@ -679,7 +682,7 @@ class AudioPlayer {
     if (preload) {
       duration = await load();
     } else {
-      await _setPlatformActive(false);
+      await _setPlatformActive(false)?.catchError((dynamic e) {});
     }
     return duration;
   }
@@ -700,8 +703,10 @@ class AudioPlayer {
       throw Exception('Must set AudioSource before loading');
     }
     if (_active) {
+      final initialSeekValues = _initialSeekValues;
+      _initialSeekValues = null;
       return await _load(await _platform, _audioSource!,
-          initialSeekValues: _initialSeekValues);
+          initialSeekValues: initialSeekValues);
     } else {
       // This will implicitly load the current audio source.
       return await _setPlatformActive(true);
@@ -736,16 +741,26 @@ class AudioPlayer {
 
   Future<Duration?> _load(AudioPlayerPlatform platform, AudioSource source,
       {_InitialSeekValues? initialSeekValues}) async {
+    final activationNumber = _activationCount;
+    void checkInterruption() {
+      if (_activationCount != activationNumber) {
+        // the platform has changed since we started loading, so abort.
+        throw PlatformException(code: 'abort', message: 'Loading interrupted');
+      }
+    }
+
     try {
       if (!kIsWeb && (source._requiresProxy || _userAgent != null)) {
         if (_proxy == null) {
           _proxy = _ProxyHttpServer();
           await _proxy!.start();
+          checkInterruption();
         }
       }
       await source._setup(this);
+      checkInterruption();
       source._shuffle(initialIndex: initialSeekValues?.index ?? 0);
-      _updateShuffleIndices();
+      _broadcastSequence();
       _durationFuture = platform
           .load(LoadRequest(
             audioSourceMessage: source._toMessage(),
@@ -754,6 +769,7 @@ class AudioPlayer {
           ))
           .then((response) => response.duration);
       final duration = await _durationFuture;
+      checkInterruption();
       _durationSubject.add(duration);
       if (platform != _platformValue) {
         // the platform has changed since we started loading, so abort.
@@ -762,6 +778,7 @@ class AudioPlayer {
       // Wait for loading state to pass.
       await processingStateStream
           .firstWhere((state) => state != ProcessingState.loading);
+      checkInterruption();
       return duration;
     } on PlatformException catch (e) {
       try {
@@ -892,7 +909,7 @@ class AudioPlayer {
   /// decoders alive so that the app can quickly resume audio playback.
   Future<void> stop() async {
     if (_disposed) return;
-    final future = _setPlatformActive(false);
+    final future = _setPlatformActive(false)?.catchError((dynamic e) {});
 
     _playInterrupted = false;
     // Update local state immediately so that queries aren't surprised.
@@ -1110,44 +1127,53 @@ class AudioPlayer {
       {Completer<void>? playCompleter, bool force = false}) {
     if (_disposed) return null;
     if (!force && (active == _active)) return _durationFuture;
+    _platformLoading = active;
+
+    // Warning! Tricky async code lies ahead.
+    // (This should definitely be made less tricky)
+    // This method itself is not asynchronous, and guarantees that _platform
+    // will be set in this cycle to a Future. The platform returned by that
+    // future takes time to initialise and so we need to handle the case where
+    // that initialisation was interrupted by another call to
+    // _setPlatformActive.
+
+    // Store the current activation sequence number. activationNumber should
+    // equal _activationCount for the duration of this call, unless it is
+    // interrupted by another simultaneous call.
+    final activationNumber = ++_activationCount;
+
+    /// Tells whether we've been interrupted.
+    bool wasInterrupted() => _activationCount != activationNumber;
+
+    final durationCompleter = Completer<Duration?>();
+
+    // Checks if we were interrupted and aborts the current activation. If we
+    // are interrupted, there are two cases:
+    // 1. If we were activating the native platform, abort with an exception.
+    // 2. If we were activating the idle dummy, abort silently.
+    //
+    // We should call this after each awaited call since those are opportunities
+    // for other coroutines to run and interrupt this one.
+    bool checkInterruption() {
+      // No interruption.
+      if (!wasInterrupted()) return false;
+      // An interruption that we can ignore
+      if (!active) return true;
+      // An interruption that should throw
+      final e =
+          PlatformException(code: 'abort', message: 'Loading interrupted');
+      durationCompleter.completeError(e);
+      throw e;
+    }
+
     // This method updates _active and _platform before yielding to the next
     // task in the event loop.
     _active = active;
     final position = this.position;
     final currentIndex = this.currentIndex;
     final audioSource = _audioSource;
-    final durationCompleter = Completer<Duration?>();
 
-    Future<AudioPlayerPlatform> setPlatform() async {
-      _playbackEventSubscription?.cancel();
-      _playerDataSubscription?.cancel();
-      if (!force) {
-        final oldPlatform = _platformValue!;
-        if (oldPlatform != _idlePlatform) {
-          await _disposePlatform(oldPlatform);
-        }
-      }
-      if (_disposed) return _platform;
-      // During initialisation, we must only use this platform reference in case
-      // _platform is updated again during initialisation.
-      final platform = active
-          ? await (_nativePlatform =
-              JustAudioPlatform.instance.init(InitRequest(
-              id: _id,
-              audioLoadConfiguration: _audioLoadConfiguration?._toMessage(),
-              androidAudioEffects: (_isAndroid() || _isUnitTest())
-                  ? _audioPipeline.androidAudioEffects
-                      .map((audioEffect) => audioEffect._toMessage())
-                      .toList()
-                  : [],
-              darwinAudioEffects: (_isDarwin() || _isUnitTest())
-                  ? _audioPipeline.darwinAudioEffects
-                      .map((audioEffect) => audioEffect._toMessage())
-                      .toList()
-                  : [],
-            )))
-          : _idlePlatform!;
-      _platformValue = platform;
+    void subscribeToEvents(AudioPlayerPlatform platform) {
       _playerDataSubscription =
           platform.playerDataMessageStream.listen((message) {
         if (message.playing != null && message.playing != playing) {
@@ -1181,9 +1207,16 @@ class AudioPlayer {
             sequence![index].duration = duration;
           }
         }
+        if (_platformLoading &&
+            message.processingState != ProcessingStateMessage.idle) {
+          _platformLoading = false;
+        }
         final playbackEvent = PlaybackEvent(
-          processingState:
-              ProcessingState.values[message.processingState.index],
+          // The platform may emit an idle state while it's starting up which we
+          // override here.
+          processingState: _platformLoading
+              ? ProcessingState.loading
+              : ProcessingState.values[message.processingState.index],
           updateTime: message.updateTime,
           updatePosition: message.updatePosition,
           bufferedPosition: message.bufferedPosition,
@@ -1206,11 +1239,53 @@ class AudioPlayer {
         if (_playbackEvent.processingState !=
                 oldPlaybackEvent.processingState &&
             _playbackEvent.processingState == ProcessingState.idle) {
-          _setPlatformActive(false);
+          _setPlatformActive(false)?.catchError((dynamic e) {});
         }
       }, onError: _playbackEventSubject.addError);
+    }
+
+    Future<AudioPlayerPlatform> setPlatform() async {
+      _playbackEventSubscription?.cancel();
+      _playerDataSubscription?.cancel();
+      if (!force) {
+        final oldPlatform = _platformValue!;
+        if (!(oldPlatform is _IdleAudioPlayer)) {
+          await _disposePlatform(oldPlatform);
+        }
+      }
+      if (_disposed) return _platform;
+      // During initialisation, we must only use this platform reference in case
+      // _platform is updated again during initialisation.
+      final platform = active
+          ? await (_nativePlatform =
+              JustAudioPlatform.instance.init(InitRequest(
+              id: _id,
+              audioLoadConfiguration: _audioLoadConfiguration?._toMessage(),
+              androidAudioEffects: (_isAndroid() || _isUnitTest())
+                  ? _audioPipeline.androidAudioEffects
+                      .map((audioEffect) => audioEffect._toMessage())
+                      .toList()
+                  : [],
+              darwinAudioEffects: (_isDarwin() || _isUnitTest())
+                  ? _audioPipeline.darwinAudioEffects
+                      .map((audioEffect) => audioEffect._toMessage())
+                      .toList()
+                  : [],
+            )))
+          : (_idlePlatform =
+              _IdleAudioPlayer(id: _id, sequenceStream: sequenceStream));
+      if (checkInterruption()) return platform;
+
+      _platformValue = platform;
 
       if (active) {
+        if (audioSource != null) {
+          _playbackEventSubject.add(_playbackEvent = _playbackEvent.copyWith(
+            updatePosition: position,
+            processingState: ProcessingState.loading,
+          ));
+        }
+
         final automaticallyWaitsToMinimizeStalling =
             this.automaticallyWaitsToMinimizeStalling;
         final playing = this.playing;
@@ -1219,12 +1294,14 @@ class AudioPlayer {
         if (_isAndroid() || _isUnitTest()) {
           if (_androidApplyAudioAttributes) {
             final audioSession = await AudioSession.instance;
+            if (checkInterruption()) return platform;
             _androidAudioAttributes ??=
                 audioSession.configuration?.androidAudioAttributes;
           }
           if (_androidAudioAttributes != null) {
             await _internalSetAndroidAudioAttributes(
                 platform, _androidAudioAttributes!);
+            if (checkInterruption()) return platform;
           }
         }
         if (!automaticallyWaitsToMinimizeStalling) {
@@ -1232,38 +1309,51 @@ class AudioPlayer {
           await platform.setAutomaticallyWaitsToMinimizeStalling(
               SetAutomaticallyWaitsToMinimizeStallingRequest(
                   enabled: automaticallyWaitsToMinimizeStalling));
+          if (checkInterruption()) return platform;
         }
         await platform.setVolume(SetVolumeRequest(volume: volume));
+        if (checkInterruption()) return platform;
         await platform.setSpeed(SetSpeedRequest(speed: speed));
+        if (checkInterruption()) return platform;
         try {
           await platform.setPitch(SetPitchRequest(pitch: pitch));
         } catch (e) {
           // setPitch not supported on this platform.
         }
+        if (checkInterruption()) return platform;
         try {
           await platform.setSkipSilence(
               SetSkipSilenceRequest(enabled: skipSilenceEnabled));
         } catch (e) {
           // setSkipSilence not supported on this platform.
         }
+        if (checkInterruption()) return platform;
         await platform.setLoopMode(SetLoopModeRequest(
             loopMode: LoopModeMessage.values[loopMode.index]));
+        if (checkInterruption()) return platform;
         await platform.setShuffleMode(SetShuffleModeRequest(
             shuffleMode: shuffleModeEnabled
                 ? ShuffleModeMessage.all
                 : ShuffleModeMessage.none));
+        if (checkInterruption()) return platform;
         if (playing) {
           _sendPlayRequest(platform, playCompleter);
         }
       }
+
+      subscribeToEvents(platform);
+
       if (audioSource != null) {
         try {
+          final initialSeekValues = _initialSeekValues ??
+              _InitialSeekValues(position: position, index: currentIndex);
+          _initialSeekValues = null;
           final duration = await _load(platform, _audioSource!,
-              initialSeekValues: _initialSeekValues ??
-                  _InitialSeekValues(position: position, index: currentIndex));
+              initialSeekValues: initialSeekValues);
+          if (checkInterruption()) return platform;
           durationCompleter.complete(duration);
         } catch (e, stackTrace) {
-          _setPlatformActive(false)?.catchError((dynamic e) {});
+          await _setPlatformActive(false)?.catchError((dynamic e) {});
           durationCompleter.completeError(e, stackTrace);
         }
       } else {
@@ -1276,12 +1366,13 @@ class AudioPlayer {
     Future<void> initAudioEffects() async {
       for (var audioEffect in _audioPipeline._audioEffects) {
         await audioEffect._activate();
+        if (checkInterruption()) return;
       }
     }
 
     _platform = setPlatform();
     if (_active) {
-      initAudioEffects();
+      initAudioEffects().catchError((dynamic e) {});
     }
     return durationCompleter.future;
   }
@@ -1316,7 +1407,7 @@ class AudioPlayer {
 /// Captures the details of any error accessing, loading or playing an audio
 /// source, including an invalid or inaccessible URL, or an audio encoding that
 /// could not be understood.
-class PlayerException {
+class PlayerException implements Exception {
   /// On iOS and macOS, maps to `NSError.code`. On Android, maps to
   /// `ExoPlaybackException.type`. On Web, maps to `MediaError.code`.
   final int code;
@@ -1334,7 +1425,7 @@ class PlayerException {
 
 /// An error that occurs when one operation on the player has been interrupted
 /// (e.g. by another simultaneous operation).
-class PlayerInterruptedException {
+class PlayerInterruptedException implements Exception {
   final String? message;
 
   PlayerInterruptedException(this.message);
@@ -1470,10 +1561,11 @@ class PlayerState {
   String toString() => 'playing=$playing,processingState=$processingState';
 
   @override
-  int get hashCode => toString().hashCode;
+  int get hashCode => hashValues(playing, processingState);
 
   @override
-  bool operator ==(dynamic other) =>
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
       other is PlayerState &&
       other.playing == playing &&
       other.processingState == processingState;
@@ -1494,11 +1586,14 @@ class IcyInfo {
   String toString() => 'title=$title,url=$url';
 
   @override
-  int get hashCode => toString().hashCode;
+  int get hashCode => hashValues(title, url);
 
   @override
-  bool operator ==(dynamic other) =>
-      other is IcyInfo && other.toString() == toString();
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is IcyInfo &&
+      other.title == title &&
+      other.url == url;
 }
 
 class IcyHeaders {
@@ -1535,8 +1630,15 @@ class IcyHeaders {
   int get hashCode => toString().hashCode;
 
   @override
-  bool operator ==(dynamic other) =>
-      other is IcyHeaders && other.toString() == toString();
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is IcyHeaders &&
+      other.bitrate == bitrate &&
+      other.genre == genre &&
+      other.name == name &&
+      other.metadataInterval == metadataInterval &&
+      other.url == url &&
+      other.isPublic == isPublic;
 }
 
 class IcyMetadata {
@@ -1553,11 +1655,14 @@ class IcyMetadata {
   IcyMetadata({required this.info, required this.headers});
 
   @override
-  int get hashCode => info.hashCode ^ headers.hashCode;
+  int get hashCode => hashValues(info, headers);
 
   @override
-  bool operator ==(dynamic other) =>
-      other is IcyMetadata && other.info == info && other.headers == headers;
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is IcyMetadata &&
+      other.info == info &&
+      other.headers == headers;
 }
 
 /// Encapsulates the [sequence] and [currentIndex] state and ensures
@@ -1937,7 +2042,10 @@ abstract class AudioSource {
   int get hashCode => _id.hashCode;
 
   @override
-  bool operator ==(dynamic other) => other is AudioSource && other._id == _id;
+  bool operator ==(Object other) =>
+      other.runtimeType == runtimeType &&
+      other is AudioSource &&
+      other._id == _id;
 }
 
 /// An [AudioSource] that can appear in a sequence.
@@ -2625,13 +2733,15 @@ class LockCachingAudioSource extends StreamAudioSource {
       for (var cacheResponse in inProgressResponses) {
         if (_progress >= cacheResponse.end) {
           // We've received enough data to fulfill the byte range request.
-          cacheResponse.controller.add(
-              data.sublist(0, data.length - (_progress - cacheResponse.end)));
+          final subEnd = min(data.length,
+              max(0, data.length - (_progress - cacheResponse.end)));
+          cacheResponse.controller.add(data.sublist(0, subEnd));
           cacheResponse.controller.close();
         } else {
           cacheResponse.controller.add(data);
         }
       }
+      inProgressResponses.removeWhere((element) => element.controller.isClosed);
       if (_requests.isEmpty) return;
       // Prevent further data coming from the HTTP source until we have set up
       // an entry in inProgressResponses to continue receiving live HTTP data.
@@ -3357,6 +3467,12 @@ class AndroidEqualizerBand {
     }
   }
 
+  /// Restores the gain after reactivating.
+  Future<void> _restore() async {
+    await (await _player._platform).androidEqualizerBandSetGain(
+        AndroidEqualizerBandSetGainRequest(bandIndex: index, gain: gain));
+  }
+
   static AndroidEqualizerBand _fromMessage(
           AudioPlayer player, AndroidEqualizerBandMessage message) =>
       AndroidEqualizerBand._(
@@ -3386,6 +3502,13 @@ class AndroidEqualizerParameters {
     required this.bands,
   });
 
+  /// Restore platform state after reactivating.
+  Future<void> _restore() async {
+    for (var band in bands) {
+      await band._restore();
+    }
+  }
+
   static AndroidEqualizerParameters _fromMessage(
           AudioPlayer player, AndroidEqualizerParametersMessage message) =>
       AndroidEqualizerParameters(
@@ -3411,7 +3534,10 @@ class AndroidEqualizer extends AudioEffect with AndroidAudioEffect {
   @override
   Future<void> _activate() async {
     await super._activate();
-    if (_parametersCompleter.isCompleted) return;
+    if (_parametersCompleter.isCompleted) {
+      await (await parameters)._restore();
+      return;
+    }
     final response = await (await _player!._platform)
         .androidEqualizerGetParameters(AndroidEqualizerGetParametersRequest());
     _parameters =