+++ just_audio_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.2
+
+* Fix bug loading audio after concat delete/insert.
+
+## 0.4.1
+
+* Remember position after stopping.
+
 ## 0.4.0
 
 * Upgrade to platform interface 4.0.0